requirements.txt
beautifulsoup4==4.6.3
certifi==2018.10.15
chardet==3.0.4
click==7.0
colorama==0.4.0 ; sys_platform == 'win32'
dash-core-components==0.38.0
dash-html-components==0.13.2
dash-renderer==0.14.3
dash-table==3.1.5
dash==0.29.0
dataclasses==0.6
decorator==4.3.0
defopt==5.0.0
docutils==0.14
flask-compress==1.4.0
flask==1.0.2
future==0.17.1
geocoder==1.38.1
geographiclib==1.49
geopy==1.17.0
idna==2.7
ipython-genutils==0.2.0
itsdangerous==1.1.0
jinja2==2.10
jsonschema==2.6.0
jupyter-core==4.4.0
lxml==4.2.5
markupsafe==1.1.0
mypy-extensions==0.4.1
nbformat==4.4.0
numpy==1.15.4
pillow==5.3.0
plotly==3.4.1
pockets==0.7.2
pony==0.7.6
python-dateutil==2.7.5
pytz==2018.7
ratelim==0.1.6
requests==2.20.1
retrying==1.3.3
robobrowser==0.5.3
singleton-decorator==1.0.0
six==1.11.0
sphinxcontrib-napoleon==0.7
tflearn==0.3.2
traitlets==4.3.2
typing-inspect==0.3.1
urllib3==1.24.1
werkzeug==0.14.1
